Problem: can't generate a new plugin template config
Solution: add command line argument for generating a config

This patch changes the name of the boostrap.py script to plugin-template. The plugin writer is now
requred to generate a template_config.yml before bootstrapping a new plugin.

Adds a --generate-config argument that generates a template_config.yml inside the plugin's root
directory. This config is then going to be used to generate Travis configuration and
scripts.

Plugin names can now be anything. The user must specify a plugin-app-label when generating a config.

This patch removes all the command line arguments that are now represented in the template config.

The settings in the template_config.yml no longer have exclude_ prefix.

The location of where the 'app' and 'tests' directories get created inside a new plugin has been
updated.

The documentation was updated to reflect the new workflow.

re: #4082
https://pulp.plan.io/issues/4082