Travis: Build on tag uses PyPI modules
When building a release on a tag, the Travis build should ensure that the
plugin works with released versions of the pulpcore and pulpcore-plugin
modules. This should catch issues like https://pulp.plan.io/issues/5509.

If the Travis build is for a tag, do not use the master versions of
these modules, but install from PyPI instead.

ref 4386
https://pulp.plan.io/issues/4386