Task: plugin-template needs to let plugins add new Travis services
Solution: Create a variable list of services that get appended to
the normal list of services.

This approach makes things easy for plugin devs, and prevents them
from overlooking adding new services that the rest of the template
needs, and running into CI failure because of it.

The only downside is that we might have plugin devs might specify
service foo, but it later becomes a fixed service, and gets listed 2x.
Duplicates might break Travis, but we could also deduplicate it in
Python logic.

fixes: #6150