Moves version to PulpPluginAppConfig subclass
Starting with pulpcore==3.9, it's expected for the version of a plugin
to be set as the `version` attribute on the `PulpPluginAppConfig`
subclass which defines the plugin.

This also updates the bumpversion config to set versions in the right
places, and the release script to now parse those locations correctly.

re #7943