Don't publish docs to the root if not the latest version
This patch updates the script that publishes docs to pulpproject.org. The version of the docs being
published is compared to the version of the package on PyPI. If the latest version of the package on
PyPI is equal to the version of the docs being published, the site root is updated with the docs.
Otherwise the docs are published to 3.y and 3.y.z directories.

re: #7766
https://pulp.plan.io/issues/7766