Add a ready-to-ship check that depends all jobs
This is a dummy job that depends on all other test runs, so declaring a
single required job in the branch protection rules is sufficient.
Also this allows to add and remove scenarios to/from our CI via a pull
request. And we can use the same rules for all release branches having
different scenarios.

[noissue]