Migrate to bump-my-version
bump-my-version is the selected heir of the discontinued bump2version
that was the most popular fork of bumpversion.
It is not a complete drop-in replacement, since most of the command
structure changed. So all the invocations here needed to be adjusted.
However this new tool can (and probably should over the course of time)
be configured in the now famous pyproject.toml file.